+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # JEC Commons Project: Update Release Notes
 
+<a name="jec-commons-1.0.1"></a>
+## **1.0.1** (2017-05-17)
+
+### Bug Fixes
+
+### Features
+
+- **GuidGenerator:** a helper class that provides a common API for generating Globally Unique Identifiers (GUID) in JEC modules.
+- **Singleton:** a new interface that can be used for detecting singleton issues due to the [Node.js Module Caching Caveats](https://nodejs.org/api/modules.html#modules_module_caching_caveats)
+
 <a name="jec-commons-1.0.0"></a>
-## **1.0.0** (2017-05-14)
+## **1.0.0** (2017-05-16)
 
 ### Bug Fixes